Deleting likes based on types

+++ routes/authors.rb
@@ -29,7 +29,13 @@ namespace '/api/v1/authors' do
     end
 
     delete '/destroy/:id' do |id|
-        Author.destroy_by(id: id)
+        if(Author.exists?(id))
+            Author.destroy_by(id: id)
+            Like.delete_by("ref_type = ? AND ref_id = ?", "author", id)
+            halt(200, {msg: "Deleted Author id = #{id}"}.to_json)
+          else
+            halt(200, {msg: "Author with id #{id} not found"}.to_json)
+        end
         halt(200, "Deleted author id = #{id}")
 
         rescue Exception => e
